refactor: change default floating number type to double, closes #4

+++ RCaron/Parsing/TokenReader.cs
@@ -119,7 +119,7 @@ public class TokenReader
             if (isDecimal)
                 return new ConstToken(TokenType.DecimalNumber,
                     (initialPosition, _position),
-                    Decimal.Parse(_code[initialPosition.._position], CultureInfo.InvariantCulture));
+                    Double.Parse(_code[initialPosition.._position], CultureInfo.InvariantCulture));
             return new ConstToken(TokenType.Number, (initialPosition, _position),
                 Int64.Parse(_code[initialPosition.._position], CultureInfo.InvariantCulture));
         }
